Widen Amount from int to long
int caps a single payment at 2,147,483,647 minor units (~£21.4M). long is
the conventional type for minor-unit amounts and removes the ceiling as a
consideration entirely.

Carried through the request, both responses, Payment, the service request
and the bank wire contract, so no narrowing happens anywhere in the path.

+++ src/PaymentGateway.Application/Models/AcquiringBankPaymentRequest.cs
@@ -5,5 +5,5 @@ public record AcquiringBankPaymentRequest(
     int ExpiryMonth,
     int ExpiryYear,
     string Currency,
-    int Amount,
+    long Amount,
     string Cvv);
